Revert "updated: prep for new updater (#31695)"
This reverts commit b93f6ce4f6fd34d33f990e86bde22b5cec49f2da.
